corrected share as image PopupMenuItem value to 1

+++ lib/util/gen_pdf.dart
@@ -1,14 +1,18 @@
 // Core
 import 'dart:io';
 import 'dart:typed_data';
+import 'dart:ui';
 
 // External
-import 'package:open_file/open_file.dart';
+// import 'package:open_file/open_file.dart';
 import 'package:pdf/pdf.dart';
 import 'package:pdf/widgets.dart' as pw;
 import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:share_plus/share_plus.dart';
+import 'package:pdf_render/pdf_render.dart' as render;
+import 'package:image/image.dart' as imglib;
+import 'package:pdf_image_renderer/pdf_image_renderer.dart';
 
 // Internal
 import '../models/transaction_record.dart';
@@ -99,6 +103,97 @@ Future<File> savePDF(TransactionRecord txnRecord) async {
 Future<void> sharePDF(TransactionRecord txnRecord) async {
   final File receiptFilePDF = await savePDF(txnRecord);
   // OpenFile.open(receiptFilePDF.path);
-  Share.shareFiles([receiptFilePDF.path], subject: 'Receipt from MeterPay.NG')
-      .whenComplete(() => receiptFilePDF.delete(recursive: true));
+  final status = await Permission.storage.status;
+  if (status.isGranted) {
+    Share.shareFiles([receiptFilePDF.path], subject: 'Receipt from MeterPay.NG')
+        .whenComplete(() => receiptFilePDF.delete(recursive: true));
+  }
+}
+
+// Future<void> shareImage() async {
+
+// final doc = await PdfDocument.openFile('abc.pdf');
+// final pages = doc.pageCount;
+// List<imglib.Image> images = [];
+
+// // get images from all the pages
+// for (int i = 1; i <= pages; i++) {
+//   var page = await doc.getPage(i);
+//   var imgPDF = await page.render();
+//   var img = await imgPDF.createImageDetached();
+//   var imgBytes = await img.toByteData(format: ImageByteFormat.png);
+//   var libImage = imglib.decodeImage(imgBytes.buffer
+//       .asUint8List(imgBytes.offsetInBytes, imgBytes.lengthInBytes));
+//   images.add(libImage);
+// }
+
+// // stitch images
+// int totalHeight = 0;
+// images.forEach((e) {
+//   totalHeight += e.height;
+// });
+// int totalWidth = 0;
+// images.forEach((element) {
+//   totalWidth = totalWidth < element.width ? element.width : totalWidth;
+// });
+// final mergedImage = imglib.Image(totalWidth, totalHeight);
+// int mergedHeight = 0;
+// images.forEach((element) {
+//   imglib.copyInto(mergedImage, element, dstX: 0, dstY: mergedHeight, blend: false);
+//   mergedHeight += element.height;
+// });
+// }
+
+// Future<render.PdfPageImage> _renderPage(render.PdfDocument document, int pageNumber) async {
+//     final page = await document.getPage(pageNumber);
+//     final pageImage = await page.render(
+//       width: page.width * 2,
+//       height: page.height * 2,
+//       format: PdfPageFormat.PNG,
+//     );
+//     await page.close();
+//     return pageImage;
+//   }
+
+void shareRenderedPDFImage(TransactionRecord txnRecord, Color color) async {
+  final File receiptFilePDF = await savePDF(txnRecord);
+  final String rootPath = (await getExternalStorageDirectory())!.path;
+  final String receiptPDFFilePath = '$rootPath/${txnRecord.txnReference}.pdf';
+
+  final pdf = PdfImageRendererPdf(path: receiptPDFFilePath);
+
+  await pdf.open();
+
+  await pdf.openPage(pageIndex: 0);
+
+  final size = await pdf.getPageSize(pageIndex: 0);
+
+  final Uint8List? image = await pdf.renderPage(
+    pageIndex: 0,
+    x: 0,
+    y: 0,
+    width: size.width,
+    height: size.height,
+    scale: 1,
+    background: color,
+  );
+
+  await pdf.closePage(pageIndex: 0);
+
+  pdf.close();
+
+  final File receiptFileImage = File('$rootPath/${txnRecord.txnReference}.png');
+
+  if (image != null) {
+    receiptFileImage.writeAsBytes(image);
+    final status = await Permission.storage.status;
+    if (status.isGranted) {
+      Share.shareFiles([receiptFileImage.path],
+              subject: 'Receipt from MeterPay.NG')
+          .whenComplete(() {
+        receiptFilePDF.delete();
+        receiptFileImage.delete();
+      });
+    }
+  }
 }
